moved task_model.py code to corva/models/task.py
