+++ corva/app/context.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-from corva.app.task_model import TaskData
+from corva.models.task import TaskData
 from corva.event.event import Event
 from corva.state.redis_state import RedisState
